[mtl] move writer into a separate module
